+++ src/PSRule/Host/HostHelper.cs
@@ -181,7 +181,7 @@ namespace PSRule.Host
             var result = new Collection<ILanguageBlock>();
             var d = new DeserializerBuilder()
                 .IgnoreUnmatchedProperties()
-                .WithNamingConvention(new CamelCaseNamingConvention())
+                .WithNamingConvention(CamelCaseNamingConvention.Instance)
                 .WithTypeConverter(new FieldMapYamlTypeConverter())
                 .WithNodeDeserializer(
                     inner => new LanguageBlockDeserializer(inner),
@@ -204,9 +204,8 @@ namespace PSRule.Host
                         using (var reader = new StreamReader(file.Path))
                         {
                             var parser = new YamlDotNet.Core.Parser(reader);
-                            parser.Expect<StreamStart>();
-
-                            while (parser.Accept<DocumentStart>())
+                            parser.TryConsume<StreamStart>(out _);
+                            while (parser.Current is DocumentStart)
                             {
                                 var item = d.Deserialize<ResourceObject>(parser: parser);
                                 if (item == null || item.Block == null)